Remove PartitionStatsV3
